Move hamster module to the classes folder
