+++ classes/hamster_app.py
